requirements.txt
requests
beautifulsoup4
googlesearch-python
duckduckgo-search
